move compute aria to dom
