+++ addon-test-support/dom/compute-aria.js
