+++ addon-test-support/finders.js
@@ -1,6 +1,6 @@
 import AmbiguousLabel from './errors/ambiguous-label';
 import MissingObject from './errors/missing-object';
-import computeAria from './utils/compute-aria';
+import computeAria from './dom/compute-aria';
 import { buttonQuery, formControlQuery } from './dom/selectors'
 import { findAll } from "@ember/test-helpers";
